parameterize test for check media logging to run directory
